crear alertas  para el boton de salir

+++ static/js/dashboard/index.js
@@ -19,9 +19,23 @@ function salir() {
   const buttonSalir = document.querySelector('.buttonSalir')
   buttonSalir.addEventListener('click', (e) => {
     e.preventDefault()
-    window.location.href = '#login'
-    // borra la sesión activa
-    localStorage.setItem('sesionActive', JSON.stringify(''))
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Cerrar sesión!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        // borra la sesión activa
+        localStorage.setItem('sesionActive', JSON.stringify(''))
+        window.location.href = '#login'
+
+      }
+
+    })
 
   })
 }
